Improve specs; implement missing features
