+++ spec/lib/mail_style_spec.rb
@@ -1,172 +1,27 @@
 require 'spec_helper'
 
-describe MailStyle, ".inline_css(css, html)" do
-  def use_css(css); @css = css; end
-  def rendering(html, options = {})
-    Nokogiri::HTML.parse MailStyle::Inlining.new(@css || "", html, {:host => options[:host] || 'example.com'}).execute
-  end
-
-  describe "inlining styles" do
-    it "should inline simple attributes" do
-      use_css 'p { color: green }'
-      rendering('<p></p>').should have_styling('color' => 'green').at_selector('p')
-    end
-
-    it "should combine multiple selectors into one" do
-      use_css "p { color: green; }
-              .tip { float: right; }"
-      rendering('<p class="tip"></p>').should have_styling('color' => 'green', 'float' => 'right').at_selector('p')
-    end
-
-    it "should use the ones attributes with the highest specificality when conflicts arises" do
-      use_css "p { color: red; }
-              .safe { color: green; border: 1px solid black; }"
-      rendering('<p class="safe"></p>').should have_styling('color' => 'green', 'border' => '1px solid black').at_selector('p')
-    end
-
-    it "should respect !important properties" do
-      pending 'TODO'
-      use_css "a { text-decoration: underline !important; }
-               a.hard-to-spot { text-decoration: none; }"
-      rendering('<a class="hard-to-spot"></a>').should have_styling('text-decoration' => 'underline').at_selector('a')
-    end
-
-    it "should combine with already present inline styles" do
-      use_css "p { color: green }"
-      rendering('<p style="font-size: 1.1em"></p>').should have_styling('color' => 'green', 'font-size' => '1.1em').at_selector('p')
-    end
-
-    it "should not overwrite already present inline styles" do
-      pending 'TODO'
-      use_css "p { color: red }"
-      rendering('<p style="color: green"></p>').should have_styling('color' => 'green').at_selector('p')
-    end
-
-    describe "inline <style> elements" do
-      it "should be used for inlined styles" do
-        rendering(<<-HTML).should have_styling('color' => 'green', 'font-size' => '1.1em').at_selector('p')
-          <html>
-            <head>
-              <style type="text/css">p { color: green; }</style>
-            </head>
-            <body>
-              <p>Hello World</p>
-              <style type="text/css">p { font-size: 1.1em; }</style>
-            </body>
-          </html>
-        HTML
-      end
-      it "should be removed" do
-        rendering(<<-HTML).should_not have_selector('style')
-          <html>
-            <head>
-              <style type="text/css">p { color: green; }</style>
-            </head>
-            <body>
-              <style type="text/css">p { font-size: 1.1em; }</style>
-            </body>
-          </html>
-        HTML
-      end
-
-      it "should not be touched when data-immutable=true" do
-        document = rendering <<-HTML
-          <html>
-            <head>
-              <style type="text/css" data-immutable="true">p { color: red; }</style>
-            </head>
-            <body>
-              <p></p>
-            </body>
-          </html>
-        HTML
-        document.should have_selector('style[data-immutable=true]')
-        document.should_not have_styling('color' => 'red').at_selector('p')
-      end
-
-      it "should not be touched when media=print" do
-        document = rendering <<-HTML
-          <html>
-            <head>
-              <style type="text/css" media="print">p { color: red; }</style>
-            </head>
-            <body>
-              <p></p>
-            </body>
-          </html>
-        HTML
-        document.should have_selector('style[media=print]')
-        document.should_not have_styling('color' => 'red').at_selector('p')
-      end
-    end
-
-  end
-
-  describe "making urls absolute" do
-    it "should work on image sources" do
-      rendering('<img src="/images/foo.jpg" />').should have_attribute('src' => 'http://example.com/images/foo.jpg').at_selector('img')
-      rendering('<img src="../images/foo.jpg" />').should have_attribute('src' => 'http://example.com/images/foo.jpg').at_selector('img')
-      rendering('<img src="foo.jpg" />').should have_attribute('src' => 'http://example.com/foo.jpg').at_selector('img')
-    end
-
-    it "should not touch image sources that are already absolute" do
-      rendering('<img src="http://other.example.org/images/foo.jpg" />').should have_attribute('src' => 'http://other.example.org/images/foo.jpg').at_selector('img')
-    end
-
-    it "should work on inlined style attributes" do
-      pending 'TODO'
-      rendering('<p style="background-image: url(/paper.png)"></p>').should  have_styling('background-image' => 'url(http://example.com/paper.jpg)').at_selector('p')
-      rendering('<p style="background-image: url(&quot;/paper.png&quot;)"></p>').should  have_styling('background-image' => 'url("http://example.com/paper.jpg")').at_selector('p')
-    end
-
-    it "should work on external style declarations" do
-      use_css "p { background-image: url(/paper.png); }
-               table { background-image: url('/paper.png'); }
-               div { background-image: url(\"/paper.png\"); }"
-      rendering('<p></p>').should have_styling('background-image' => 'url(http://example.com/paper.png)').at_selector('p')
-      rendering('<table></table>').should have_styling('background-image' => "url('http://example.com/paper.png')").at_selector('table')
-      rendering('<div></div>').should have_styling('background-image' => 'url("http://example.com/paper.png")').at_selector('div')
-    end
-
-    it "should not touch style urls that are already absolute" do
-      external_url = 'url(http://other.example.org/paper.png)'
-      use_css "p { background-image: #{external_url}; }"
-      rendering('<p></p>').should have_styling('background-image' => external_url).at_selector('p')
-      rendering(%(<div style="background-image: #{external_url}"></div>)).should have_styling('background-image' => external_url).at_selector('div')
+describe MailStyle do
+  describe ".inline_css" do
+    it "should create an instance of MailStyle::Inlining and executing it" do
+      MailStyle::Inlining.should_receive(:new).with('attri', 'butes').and_return(double('inlining', :execute => 'html'))
+      MailStyle.inline_css('attri', 'butes').should == 'html'
     end
   end
 
-  describe "inserting tags" do
-    it "should insert a doctype if not present" do
-      rendering('<html><body></body></html>').to_xml.should include('<!DOCTYPE ')
-      rendering('<!DOCTYPE html><html><body></body></html>').to_xml.should_not match(/(DOCTYPE.*?){2}/)
+  describe ".load_css(root, targets)" do
+    let(:fixtures_root) { Pathname.new(__FILE__).dirname.join('..', 'fixtures') }
+
+    it "should load files matching the target names under root/public/stylesheets" do
+      MailStyle.load_css(fixtures_root, ['foo']).should == 'contents of foo'
     end
 
-    it "should set xmlns of <html> to that of XHTML" do
-      rendering('<html><body></body></html>').should have_selector('html[xmlns="http://www.w3.org/1999/xhtml"]')
+    it "should load files in order and join them with a newline" do
+      MailStyle.load_css(fixtures_root, %w[foo bar]).should == "contents of foo\ncontents of bar"
+      MailStyle.load_css(fixtures_root, %w[bar foo]).should == "contents of bar\ncontents of foo"
     end
 
-    it "should insert basic html structure if not present" do
-      rendering('<h1>Hey!</h1>').should have_selector('html > head + body > h1')
-    end
-
-    it "should insert <head> if not present" do
-      rendering('<html><body></body></html>').should have_selector('html > head + body')
-    end
-
-    it "should insert meta tag describing content-type" do
-      rendering('<html><head></head><body></body></html>').should have_selector('head meta[http-equiv="Content-Type"][content="text/html; charset=utf-8"]')
-    end
-
-    it "should not insert duplicate meta tags describing content-type" do
-      pending 'TODO'
-      rendering(<<-HTML).should have_selector('meta').limit(1)
-      <html>
-        <head>
-          <meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1" />
-        </head>
-      </html>
-      HTML
+    it "should raise a MailStyle::CSSFileNotFound error when a css file could not be found" do
+      expect { MailStyle.load_css(fixtures_root, ['not_here']) }.to raise_error(MailStyle::CSSFileNotFound, /not_here/)
     end
   end
 end
